Compile literal and function invocation expressions
- Make expressions return their values in RAX by convention
- Improve comments
- Improve labels numbering in generated ASM
- Change used DoubleValue underlying type from long double to double
- Move getSize method from Variable to Type as static
- Add getGeneratedAsm() to BuiltInFunScope
- Add helper methods in Compiler
- Prepare for syscall built-in

+++ src/lexer/LexerLine.cpp
@@ -526,7 +526,7 @@ void LexerLine::getIntNumberToken(
 
 void LexerLine::getDoubleNumberToken(std::wstring* number){
     try{
-        auto doubleNum=std::stold(*number);
+        auto doubleNum=std::stod(*number);
         *number=std::to_wstring(doubleNum);
     }
     catch(std::out_of_range e){
